Hot Fix

+++ 0x0D-preprocessor/4-sum.h
@@ -1,6 +1,6 @@
 #ifndef SUM_H
 #define SUM_H
 
-#define SUM(x, y) ((long int)x + (long int)y)
+#define SUM(x, y) (x + y)
 
 #endif
